add logic and move tests to pytest
